requirements.txt
python-jose[cryptography]
sanic
sanic-ext